+++ tests/unit/canopy_model/test_canopy.py
@@ -1,0 +1,1 @@
+"""test the canopy object in canopy.py initialises as expected."""
