+++ tests/unit/canopy_model/test_canopy_functions.py
@@ -1,0 +1,1 @@
+"""test the functions in canopy_functions.py."""
