+++ tests/unit/canopy_model/test_community.py
@@ -1,0 +1,1 @@
+"""test the community object in community.py initialises as expected."""
